feature/1670 - Resolve circular dependency over buildPolyPositionModels()
